+++ src/app/orders/store/orders.effects.ts
@@ -3,7 +3,14 @@ import { Actions, createEffect, ofType } from '@ngrx/effects';
 import { catchError, map, of, switchMap } from 'rxjs';
 
 import { OrdersService } from '../services/orders.service';
-import { loadOrders, loadOrdersFailure, loadOrdersSuccess } from './orders.actions';
+import {
+  createOrder,
+  createOrderFailure,
+  createOrderSuccess,
+  loadOrders,
+  loadOrdersFailure,
+  loadOrdersSuccess,
+} from './orders.actions';
 
 @Injectable()
 export class OrdersEffects {
@@ -17,6 +24,18 @@ export class OrdersEffects {
         this.ordersService.getOrders().pipe(
           map(orders => loadOrdersSuccess({ orders })),
           catchError(error => of(loadOrdersFailure({ error: error.message })))
+        )
+      )
+    )
+  );
+
+  createOrder$ = createEffect(() =>
+    this.actions$.pipe(
+      ofType(createOrder),
+      switchMap(({ dto }) =>
+        this.ordersService.createOrder(dto).pipe(
+          map(order => createOrderSuccess({ order })),
+          catchError(error => of(createOrderFailure({ error: error.message })))
         )
       )
     )