+++ src/app/orders/store/orders.reducer.ts
@@ -1,23 +1,45 @@
 import { createReducer, on } from '@ngrx/store';
 
 import { initialOrdersState, OrdersState } from './orders.state';
-import { loadOrders, loadOrdersFailure, loadOrdersSuccess } from './orders.actions';
+import {
+  createOrder,
+  createOrderFailure,
+  createOrderSuccess,
+  loadOrders,
+  loadOrdersFailure,
+  loadOrdersSuccess,
+} from './orders.actions';
 
 export const ordersReducer = createReducer<OrdersState>(
   initialOrdersState,
   on(loadOrders, (state): OrdersState => ({
     ...state,
     loading: true,
-    error: null
+    error: null,
   })),
   on(loadOrdersSuccess, (state, { orders }): OrdersState => ({
     ...state,
     orders,
-    loading: false
+    loading: false,
   })),
   on(loadOrdersFailure, (state, { error }): OrdersState => ({
     ...state,
     loading: false,
-    error
+    error,
+  })),
+  on(createOrder, (state): OrdersState => ({
+    ...state,
+    creating: true,
+    createError: null,
+  })),
+  on(createOrderSuccess, (state, { order }): OrdersState => ({
+    ...state,
+    orders: [...state.orders, order],
+    creating: false,
+  })),
+  on(createOrderFailure, (state, { error }): OrdersState => ({
+    ...state,
+    creating: false,
+    createError: error,
   }))
 );